docs(README): introduce the type in Papyrus
    There are three types used in the Papyrus: primary key, value and
    the tag.

    The primary key and value are already defined, and docs/types.md
    introduce the definitation of the primary key and value.

+++ src/tests/test_types.py
@@ -136,7 +136,7 @@ class TestKey:
         key = Key(raw)
 
         assert key == raw
-        assert key.ktype == KeyType.BYTE
+        assert key.ktype == KeyType.WORD
 
     def test_integer_key(self, faker):
         raw = faker.pyint(min_value=32768, max_value=65535)
